+++ samples/multi-module-sample/home/src/main/java/com/kpstv/home/HomeDependency.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import javax.inject.Inject
 
 class HomeDependency @Inject constructor() {
-  fun call() {
-    Log.e("HomeDependency", "Function called")
+  fun call(tag: String) {
+    Log.e("HomeDependency:$tag", "Function called")
   }
 }