+++ samples/multi-module-sample/home/src/main/java/com/kpstv/home/di/HomeScope.kt
@@ -1,6 +1,9 @@
 package com.kpstv.home.di
 
+import dagger.hilt.android.scopes.ActivityScoped
+import dagger.hilt.migration.AliasOf
 import javax.inject.Scope
 
 @Scope
+@AliasOf(ActivityScoped::class)
 annotation class HomeScope()