+++ samples/multi-module-sample/home/src/main/java/com/kpstv/home/fragments/HomeStartFragment.kt
@@ -2,8 +2,10 @@ package com.kpstv.home.fragments
 
 import android.os.Bundle
 import android.view.View
+import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
+import androidx.appcompat.widget.Toolbar
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import com.kpstv.home.HomeDependency
@@ -11,18 +13,23 @@ import com.kpstv.home.R
 import com.kpstv.home_internal.HomeInternalFragment
 import com.kpstv.navigation.FragmentNavigator
 import com.kpstv.navigation.ValueFragment
+import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collect
 import javax.inject.Inject
 
-class HomeStartFragment @Inject constructor(
-  private val homeDependency: HomeDependency
-) : ValueFragment(R.layout.fragment_home_start) {
+@AndroidEntryPoint
+class HomeStartFragment : ValueFragment(R.layout.fragment_home_start) {
   private val viewModel by viewModels<HomeStartViewModel>()
+
+  @Inject lateinit var homeDependency: HomeDependency
 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
 
-    homeDependency.call() // constructor injected dependency
+    // don't do this, this is just for differentiating nested modules
+    requireParentFragment().requireView().findViewById<Toolbar>(R.id.toolbar).title = "Module: home"
+
+    homeDependency.call("home-start") // constructor injected dependency
 
     val stateTextView = view.findViewById<TextView>(R.id.tv_state)
     val button = view.findViewById<Button>(R.id.btn_goto)