+++ samples/multi-module-sample/home/src/main/java/com/kpstv/home/navigation/home_internal/HomeButtonClickedImpl.kt
@@ -1,12 +1,14 @@
-package com.kpstv.home
+package com.kpstv.home.navigation.home_internal
 
+import com.kpstv.home.HomeFragment
+import com.kpstv.home.di.HomeScope
 import com.kpstv.home_internal.HomeButtonClicked
 import com.kpstv.home_internal2.HomeInternal2Fragment
 import com.kpstv.navigation.FragmentNavigator
 import javax.inject.Inject
 
 class HomeButtonClickedImpl @Inject constructor(
-  private val fragment: HomeFragment
+  @HomeScope private val fragment: HomeFragment
 ) : HomeButtonClicked {
   override fun goToNext(navOptions: FragmentNavigator.NavOptions) {
     fragment.getNavigator().navigateTo(HomeInternal2Fragment::class, navOptions)